BasicUserService 작성(의존성 추가), File/JCFUserService, Service에 해당하는 File, JCF패키지삭제
